refactor: Structure packages inside the validation package
