+++ backend/src/main/java/ru/caselab/edm/backend/validation/annotations/Password.java
@@ -1,4 +1,4 @@
-package ru.caselab.edm.backend.validation.password;
+package ru.caselab.edm.backend.validation.annotations;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
