+++ backend/src/main/java/ru/caselab/edm/backend/validation/validators/PasswordMatcherValidator.java
@@ -1,7 +1,9 @@
-package ru.caselab.edm.backend.validation.password;
+package ru.caselab.edm.backend.validation.validators;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
+import ru.caselab.edm.backend.validation.annotations.PasswordMatcher;
+import ru.caselab.edm.backend.validation.interfaces.PasswordValidatable;
 
 public class PasswordMatcherValidator implements ConstraintValidator<PasswordMatcher, PasswordValidatable> {
